fix: prevent stateful regex leaking across recursive MIME-part scan
- doc(README): fixed the feature-toggle sentence
- doc(README): removed duplicate hash_date=false
- test: split into separate files
- test: refactor to shed 400 LOC
- test: updated test-fixtures
