+++ package.json
@@ -16,7 +16,8 @@
     "test": "node --test test/*.js",
     "versions": "npx npm-dep-mgr check",
     "versions:fix": "npx npm-dep-mgr update",
-    "test:coverage": "node --test --experimental-test-coverage --test-coverage-exclude=package.json --test-coverage-exclude=test/*.js test/*.js"
+    "test:coverage": "node --test --experimental-test-coverage --test-coverage-exclude=package.json --test-coverage-exclude=test/*.js test/*.js",
+    "test:coverage:lcov": "mkdir -p coverage && node --test --experimental-test-coverage --test-reporter=lcov --test-reporter-destination=coverage/lcov.info test/*.js"
   },
   "repository": {
     "type": "git",
@@ -39,8 +40,9 @@
     "haraka-plugin-spf": "^1.2.11"
   },
   "devDependencies": {
+    "address-rfc2821": "^2.2.0",
     "@haraka/eslint-config": "^2.0.4",
-    "haraka-test-fixtures": "~1.3.10",
+    "haraka-test-fixtures": "~1.4.1",
     "sinon": "^22.0.0"
   },
   "prettier": {
